cluster/health: move node-local serialization
Move serialization of node-local health state to live with the
definitions of those types.
