+++ src/v/cluster/health_monitor_types.cc
@@ -143,24 +143,6 @@ void read_and_assert_version(std::string_view type, iobuf_parser& parser) {
       T::current_version);
 }
 
-void adl<cluster::node::disk>::to(iobuf& out, cluster::node::disk&& s) {
-    serialize(out, s.current_version, s.path, s.free, s.total);
-}
-
-cluster::node::disk adl<cluster::node::disk>::from(iobuf_parser& p) {
-    read_and_assert_version<cluster::node::disk>("cluster::node::disks", p);
-
-    auto path = adl<ss::sstring>{}.from(p);
-    auto free = adl<uint64_t>{}.from(p);
-    auto total = adl<uint64_t>{}.from(p);
-
-    return cluster::node::disk{
-      .path = path,
-      .free = free,
-      .total = total,
-    };
-}
-
 void adl<cluster::node_state>::to(iobuf& out, cluster::node_state&& s) {
     serialize(out, s.current_version, s.id, s.membership_state, s.is_alive);
 }
